BUG/MEDIUM: mux-h2: Be sure to enter in demux loop even if dbuf is empty
I introduced a regression when I backported the commit df203d88
("BUG/MEDIUM: mux-h2: do not quit the demux loop before setting
END_REACHED"). In higher versions, we always enter in the demux loop even if
the demux buffer is empty, mainly to emit a trace when the buffer is
empty. While on the 2.0, we only loop if there is some data in the demux
buffer. Thus the mentionned commit has no effect on the 2.0 and does not fix
the bug it should fix.

To fix the bug, we now do the same in the 2.0. To be short, we now use a
"while (1)" instead of a "while (b_data(&h2c->dbuf))".

This fix is specific for 2.0. Thus there is no upstream commit ID. No
backport needed.